requirements.txt
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coverage==7.1.0
et-xmlfile==1.1.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
google-auth==2.12.0
google-auth-oauthlib==0.5.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
oauthlib==3.2.1
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytest==7.1.3
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.5.post1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.12
Werkzeug==2.2.2
xlwt==1.3.0
zipp==3.8.1